requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
djangorestframework==3.16.1
python-dotenv==1.2.1
flake8==7.3.0
idna==3.11
mccabe==0.7.0
mysqlclient==2.2.8
pycodestyle==2.14.0
pyflakes==3.4.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
